Handle Ctrl-{ and Ctrl-} as extended keys
Ctrl-{ and Ctrl-} don't work in tmux and they'll be sent as Ctrl-[ and Ctrl-]. This fixes it and sends Ctrl-{ and Ctrl-} as they are.

The motivation is to use these key bindings in emacs. They're bound to `paredit-backward-barf-sexp` and `paredit-forward-barf-sexp` in paredit-mode. It's annoying not to be able to use them to edit lisp code.

I can't be sure the exact implication of this change. It might break something but so far is working for me.